backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0  # 使用纯Python的MySQL驱动
python-dotenv==1.0.0